Split msnodesqlv8
