+++ lib/msnodesqlv8/request.js
@@ -2,24 +2,16 @@
 
 const msnodesql = require('msnodesqlv8')
 const debug = require('debug')('mssql:msv8')
+const BaseRequest = require('../base/request')
+const RequestError = require('../error/request-error')
+const { IDS, objectHasProperty } = require('../utils')
+const { TYPES, DECLARATIONS, declare } = require('../datatypes')
+const { PARSERS: UDT } = require('../udt')
+const Table = require('../table')
 
-const base = require('./base')
-const TYPES = require('./datatypes').TYPES
-const declare = require('./datatypes').declare
-const UDT = require('./udt').PARSERS
-const Table = require('./table')
-const DECLARATIONS = require('./datatypes').DECLARATIONS
-const ISOLATION_LEVEL = require('./isolationlevel')
-const IDS = require('./utils').IDS
-const INCREMENT = require('./utils').INCREMENT
-const objectHasProperty = require('./utils').objectHasProperty
-
-const EMPTY_BUFFER = Buffer.alloc(0)
 const JSON_COLUMN_ID = 'JSON_F52E2B61-18A1-11d1-B105-00805F49916B'
 const XML_COLUMN_ID = 'XML_F52E2B61-18A1-11d1-B105-00805F49916B'
-
-const CONNECTION_STRING_PORT = 'Driver=SQL Server Native Client 11.0;Server=#{server},#{port};Database=#{database};Uid=#{user};Pwd=#{password};Trusted_Connection=#{trusted};Encrypt=#{encrypt};'
-const CONNECTION_STRING_NAMED_INSTANCE = 'Driver=SQL Server Native Client 11.0;Server=#{server}\\#{instance};Database=#{database};Uid=#{user};Pwd=#{password};Trusted_Connection=#{trusted};Encrypt=#{encrypt};'
+const EMPTY_BUFFER = Buffer.alloc(0)
 
 const castParameter = function (value, type) {
   if (value == null) {
@@ -121,17 +113,6 @@ const createColumns = function (metadata) {
   return out
 }
 
-const isolationLevelDeclaration = function (type) {
-  switch (type) {
-    case ISOLATION_LEVEL.READ_UNCOMMITTED: return 'READ UNCOMMITTED'
-    case ISOLATION_LEVEL.READ_COMMITTED: return 'READ COMMITTED'
-    case ISOLATION_LEVEL.REPEATABLE_READ: return 'REPEATABLE READ'
-    case ISOLATION_LEVEL.SERIALIZABLE: return 'SERIALIZABLE'
-    case ISOLATION_LEVEL.SNAPSHOT: return 'SNAPSHOT'
-    default: throw new base.TransactionError('Invalid isolation level.')
-  }
-}
-
 const valueCorrection = function (value, metadata) {
   if ((metadata.sqlType === 'time') && (value != null)) {
     value.setFullYear(1970)
@@ -147,157 +128,7 @@ const valueCorrection = function (value, metadata) {
   }
 }
 
-class ConnectionPool extends base.ConnectionPool {
-  _poolCreate () {
-    return new base.Promise((resolve, reject) => {
-      let defaultConnectionString = CONNECTION_STRING_PORT
-
-      if (this.config.options.instanceName != null) {
-        defaultConnectionString = CONNECTION_STRING_NAMED_INSTANCE
-      }
-
-      if (this.config.requestTimeout == null) {
-        this.config.requestTimeout = 15000
-      }
-
-      const cfg = {
-        conn_str: this.config.connectionString || defaultConnectionString,
-        conn_timeout: (this.config.connectionTimeout || 15000) / 1000
-      }
-
-      cfg.conn_str = cfg.conn_str.replace(new RegExp('#{([^}]*)}', 'g'), (p) => {
-        const key = p.substr(2, p.length - 3)
-
-        switch (key) {
-          case 'instance':
-            return this.config.options.instanceName
-          case 'trusted':
-            return this.config.options.trustedConnection ? 'Yes' : 'No'
-          case 'encrypt':
-            return this.config.options.encrypt ? 'Yes' : 'No'
-          default:
-            return this.config[key] != null ? this.config[key] : ''
-        }
-      })
-
-      const connedtionId = INCREMENT.Connection++
-      debug('pool(%d): connection #%d created', IDS.get(this), connedtionId)
-      debug('connection(%d): establishing', connedtionId)
-
-      if (typeof this.config.beforeConnect === 'function') {
-        this.config.beforeConnect(cfg)
-      }
-
-      msnodesql.open(cfg, (err, tds) => {
-        if (err) {
-          err = new base.ConnectionError(err)
-          return reject(err)
-        }
-
-        IDS.add(tds, 'Connection', connedtionId)
-        debug('connection(%d): established', IDS.get(tds))
-        resolve(tds)
-      })
-    })
-  }
-
-  _poolValidate (tds) {
-    return tds && !tds.hasError
-  }
-
-  _poolDestroy (tds) {
-    return new base.Promise((resolve, reject) => {
-      if (!tds) {
-        resolve()
-        return
-      }
-      debug('connection(%d): destroying', IDS.get(tds))
-      tds.close(() => {
-        resolve()
-      })
-      debug('connection(%d): destroyed', IDS.get(tds))
-    })
-  }
-}
-
-class Transaction extends base.Transaction {
-  _begin (isolationLevel, callback) {
-    super._begin(isolationLevel, err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): begin', IDS.get(this))
-
-      this.parent.acquire(this, (err, connection, config) => {
-        if (err) return callback(err)
-
-        this._acquiredConnection = connection
-        this._acquiredConfig = config
-
-        const req = new Request(this)
-        req.stream = false
-        req.query(`set transaction isolation level ${isolationLevelDeclaration(this.isolationLevel)};begin tran;`, err => {
-          if (err) {
-            this.parent.release(this._acquiredConnection)
-            this._acquiredConnection = null
-            this._acquiredConfig = null
-
-            return callback(err)
-          }
-
-          debug('transaction(%d): begun', IDS.get(this))
-
-          callback(null)
-        })
-      })
-    })
-  }
-
-  _commit (callback) {
-    super._commit(err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): commit', IDS.get(this))
-
-      const req = new Request(this)
-      req.stream = false
-      req.query('commit tran', err => {
-        if (err) err = new base.TransactionError(err)
-
-        this.parent.release(this._acquiredConnection)
-        this._acquiredConnection = null
-        this._acquiredConfig = null
-
-        if (!err) debug('transaction(%d): commited', IDS.get(this))
-
-        callback(null)
-      })
-    })
-  }
-
-  _rollback (callback) {
-    super._commit(err => {
-      if (err) return callback(err)
-
-      debug('transaction(%d): rollback', IDS.get(this))
-
-      const req = new Request(this)
-      req.stream = false
-      req.query('rollback tran', err => {
-        if (err) err = new base.TransactionError(err)
-
-        this.parent.release(this._acquiredConnection)
-        this._acquiredConnection = null
-        this._acquiredConfig = null
-
-        if (!err) debug('transaction(%d): rolled back', IDS.get(this))
-
-        callback(null)
-      })
-    })
-  }
-}
-
-class Request extends base.Request {
+class Request extends BaseRequest {
   _batch (batch, callback) {
     this._isBatch = true
     this._query(batch, callback)
@@ -310,11 +141,11 @@ class Request extends base.Request {
       table._makeBulk()
 
       if (!table.name) {
-        setImmediate(callback, new base.RequestError('Table name must be specified for bulk insert.', 'ENAME'))
+        setImmediate(callback, new RequestError('Table name must be specified for bulk insert.', 'ENAME'))
       }
 
       if (table.name.charAt(0) === '@') {
-        setImmediate(callback, new base.RequestError("You can't use table variables for bulk insert.", 'ENAME'))
+        setImmediate(callback, new RequestError("You can't use table variables for bulk insert.", 'ENAME'))
       }
 
       this.parent.acquire(this, (err, connection) => {
@@ -325,7 +156,7 @@ class Request extends base.Request {
           if (this.canceled) {
             debug('request(%d): canceled', IDS.get(this))
             this.parent.release(connection)
-            return callback(new base.RequestError('Canceled.', 'ECANCEL'))
+            return callback(new RequestError('Canceled.', 'ECANCEL'))
           }
 
           const done = (err, rowCount) => {
@@ -340,7 +171,7 @@ class Request extends base.Request {
                 connection.hasError = true
               }
 
-              err = new base.RequestError(err)
+              err = new RequestError(err)
               err.code = 'EREQUEST'
             }
 
@@ -357,7 +188,7 @@ class Request extends base.Request {
             const tm = connection.tableMgr()
             return tm.bind(table.path.replace(/\[|\]/g, ''), mgr => {
               if (mgr.columns.length === 0) {
-                return done(new base.RequestError('Table was not found on the server.', 'ENAME'))
+                return done(new RequestError('Table was not found on the server.', 'ENAME'))
               }
 
               const rows = []
@@ -463,7 +294,7 @@ class Request extends base.Request {
         if (this.canceled) {
           debug('request(%d): canceled', IDS.get(this))
           this.parent.release(connection)
-          return callback(new base.RequestError('Canceled.', 'ECANCEL'))
+          return callback(new RequestError('Canceled.', 'ECANCEL'))
         }
 
         const params = []
@@ -507,7 +338,7 @@ class Request extends base.Request {
                   if (!this.stream) { recordsets[recordsets.length - 1][0] = row }
                 } catch (ex) {
                   row = null
-                  const ex2 = new base.RequestError(`Failed to parse incoming JSON. ${ex.message}`, 'EJSON')
+                  const ex2 = new RequestError(`Failed to parse incoming JSON. ${ex.message}`, 'EJSON')
 
                   if (this.stream) {
                     this.emit('error', ex2)
@@ -623,7 +454,7 @@ class Request extends base.Request {
             connection.hasError = true
           }
 
-          err = new base.RequestError(err)
+          err = new RequestError(err)
           err.code = 'EREQUEST'
           err.state = err.sqlstate
 
@@ -655,7 +486,7 @@ class Request extends base.Request {
                     if (!this.stream) { recordsets[recordsets.length - 1][0] = row }
                   } catch (ex) {
                     row = null
-                    const ex2 = new base.RequestError(`Failed to parse incoming JSON. ${ex.message}`, 'EJSON')
+                    const ex2 = new RequestError(`Failed to parse incoming JSON. ${ex.message}`, 'EJSON')
 
                     if (this.stream) {
                       this.emit('error', ex2)
@@ -808,24 +639,4 @@ class Request extends base.Request {
   }
 }
 
-module.exports = Object.assign({
-  ConnectionPool,
-  Transaction,
-  Request,
-  PreparedStatement: base.PreparedStatement
-}, base.exports)
-
-Object.defineProperty(module.exports, 'Promise', {
-  enumerable: true,
-  get: () => {
-    return base.Promise
-  },
-  set: (value) => {
-    base.Promise = value
-  }
-})
-
-base.driver.name = 'msnodesqlv8'
-base.driver.ConnectionPool = ConnectionPool
-base.driver.Transaction = Transaction
-base.driver.Request = Request
+module.exports = Request
